requirements.txt
anyio==3.6.2
anywidget==0.3.0
appnope==0.1.3
asgiref==3.7.0
asttokens==2.2.1
backcall==0.2.0
comm==0.1.3
contourpy==1.0.7
crispy==0.7.4
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
Django==4.2.1
django-google-maps==0.13.0
django-jazzmin==2.6.0
django-leaflet==0.29.0
django-map-widgets==0.4.1
executing==1.2.0
fabio==2023.4.1
fonttools==4.39.4
geographiclib==2.0
geopy==2.3.0
h5py==3.8.0
idna==3.4
ipykernel==6.23.1
ipython==8.13.2
ipywidgets==8.0.6
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
leaflet==0.0.3
mapwidget==0.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numpy==1.24.3
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2-binary==2.9.6
psygnal==0.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
python-dateutil==2.8.2
pyzmq==25.0.2
shapely==2.0.1
silx==1.1.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.0
watchfiles==0.19.0
wcwidth==0.2.6
widgetsnbextension==4.0.7
xyzservices==2023.5.0
